Make seq field in chpl_comm_bundleData_t unconditional
Previously the seq field in chpl_comm_bundleData_t was
conditionally defined based on CHPL_COMM_DEBUG. Since
chpl_comm_bundleData_t is used in the API between the application and
the runtime, this meant that both had to be compiled with
CHPL_COMM_DEBUG either set or not set. Unfortunately, the application
makefiles did not set the CHPL_COMM_DEBUG compiler flag properly
causing a wild store in the runtime. This change makes the seq field
unconditional, avoiding the dependency on CHPL_COMM_DEBUG. The field is
generally useful and the bytes saved by not including it when
CHPL_COMM_DEBUG is not set isn't worth the complexity and the danger of
a wild store.

While tracking down this bug I fixed a missing CFG_HINT_NULL in
hintVals.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>