Add deprecation warning to these tests too
I didn't pick up on the fact that there were manual applications of
the FLAG_ that I was cueing off of for my deprecation warnings in the
test system, but here they are.  Though the deprecation warning isn't
meaningful here, this also isn't a user-facing feature, so it seems
reasonable to include it in the .good file.  That said, this also
makes me curious whether there's code in the compiler that can be
removed once we remove this pragma and vectorizeOnly... I think I'd
been imagining more sharing of code between foreach and vectorizeOnly,
but if these tests are triggering this warning and others aren't...
maybe there's more dead code to remove than I'd imagined?

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>