Update CI image builder workflow
In #19425 I added a new GH Action to build a container. It used the
"on push to main" trigger, but while the new Workflow shows up at [1],
it did not run. I incorrectly assumed we could use variable
substitution in the on.paths section, but I can't get that to work.

I've also added the "on workflow_dispatch" trigger (see [2]) to
allow it to be manually run.

I've tested this in my own fork and verified:
  * a manual run will build the image and push to ghcr
  * a change to util/dockerfiles/github-ci/Dockerfile triggers a
  rebuild after merging to main

[1] https://github.com/chapel-lang/chapel/actions/workflows/build-CI-container.yml
[2] https://docs.github.com/en/actions/using-workflows/events-that-trigger-workflows#workflow_dispatch

Signed-off-by: Andrew Consroe <andrew.consroe@hpe.com>