Only declare and use mli debugging variables when debugging is enabled
As C compiler warnings get better at dead variable analysis, we're
getting warnings about variables that are only necessary when
debugging.  That is, they're always declared and set, but then
never used unless debug printing is on.  Here, we wrap their
declarations and assignments in #ifdefs to only do anything with
them if debugging is enabled.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>