Fix cross-reference to new module
Apparently in adopting the "this module is deprecated" text from the
Spawn module to make sure I adhered to the prescribed wording, I
didn't even take the time to update the name of the module that
replaces ChapelEnv — embarrassing!  Thanks Andrew, for catching this.
Even more embarrassing...  This is not the first time I've done this
this calendar year.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>