Move documentation index to sidebar and remove from bottoms of pages
While our online documentation index is a bit overwhelmingly large and
badly formatted, it still seems useful, but difficult to discover by
not being on our docs sidebar.  This PR moves it there and removes it
from other places it was mentioned.

It turns out that because of the way the index is generated by Sphinx,
it's not trivial to get it onto the sidebar, which is why I think it
was not there before, and was instead at the bottom of several
module-based pages.  Here, I took a tip from

https://stackoverflow.com/questions/25243482/how-to-add-sphinx-generated-index-to-the-sidebar-when-using-read-the-docs-theme

and wedged it in there, though to make it look proper on the web and
in the generated HTML, I had to embed more HTML than that solution
included.

I removed it from the bottoms of the other pages because it seemed
unnecessary (it's not specific to those pages in any way) and
clutter-y.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>