Throw `-DCHPL_COMM_DEBUG` when building user programs, when appropriate.
When the environment variable `CHPL_COMM_DEBUG` is defined we need to
define the C preprocessor variable of the same name when building user
programs, to adjust certain declarations in the runtime include files
that are shared with the Chapel module code.  Here, ensure that we do
so.

Support for this was inadvertently removed in #18880.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>