Update `use` statement productions in spec to allow naming operators
Update the `exclude-list` production to allow naming operators in `except`
clauses.  Update the `rename-base` production to allow naming operators in
`only` clauses.  Both of these productions are used in `import` productions, so
it also benefits from these additions.

I still want to look at the descriptions in these sections to see if there's
anything else that needs updating for operators

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>