Use chapel-github-ci container image in workflows
Enables the use of the container image at [1] in CI.yml workflows.
This avoids having to reinstall system dependencies on each
invocation. Typical load times to initialize the conatiner is about
30s. Sometimes, the apt-install step was well below this, on the order
of 10s. But in the worst case, installs could take 5+ min with a very
slow network. The speed advantage of this approach relies on the
assumption that the network for ghcr is much more reliable than that
to apt mirrors.

  [1] https://github.com/chapel-lang/chapel/pkgs/container/chapel-github-ci

Signed-off-by: Andrew Consroe <andrew.consroe@hpe.com>