Fix a compiler assert
Resolves #18218.

A loop over a recursive iterator used to trigger an assertion failure
in the compiler when the loop body contained a call to a throwing function.
This PR makes the compiler handle that case using a heuristic solution.

Adds a beefed-up reproducer from 18218 and de-futurizes a test
filed previously for 18218.

Signed-off-by: Vassily Litvinov <vasslitvinov@users.noreply.github.com>