Fixes and updates to check_paths
When check_paths was updated for the master->main change, the github
url we look for was updated but the length to remove from the token
wasn't.  Fix that (44->42).

Add a few more written-out alternatives to be ignored.

Update location and extension for the adding-a-tasking-model and
comm-layer documents, and add new special exceptions for
adding-a-launcher, SpellChecking, and RuntimeLibrary.

Add a few symbols (<, >, **) to the list to split tokens on.

---
Signed-off-by: Paul Cassella <gitgitgit@manetheren.bigw.org>