Adjust instantiatedClassname.chpl to avoid nilable class
This test originated in 942157fc18e07f025891534b8eb8a4ed9a973917 as a bug
report about class names for generic classes. This commit changes it to
no longer use a nilable class that is default initialized (since that
seems to be unrelated to the original purpose of the test).

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>