Extend the title of the section on generic compiler generated initializers
Prior to this change, the title was indistinguishable from that for concrete
compiler-generated initializers.  This was a problem when both links were
side-by-side, as is the case in the first paragraph of the concrete
compiler-generated initializer section, making it look like it linked to itself.

With this change, the section is now distinguishable.

This impacts a total of six links to that section.  Since the references all
indicate that the generic compiler-generated initializer is desired, I think it
is fine for their reference to be updated, but it would be reasonable to double
check me.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>