Add return types to documentation for array.size, .sizeAs, .shape
While doing a final check for URLs in CHANGES.md, I realized that
I hadn't added links to these routines for domains and arrays,
but then balked at putting the array links in because they didn't
declare their return types (so didn't illustrate the point that
I was trying to make).

Here, I add the return types, but break the .shape routine into two
versions to match what we do in domains, no doc-ing the version for
associative domains for consistency since it's no-doc'd there as well.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>