Limit I/O size between channel buffer and underlying file
Do I/O between a channel buffer and the underlying file in fixed-size
I/Os, rather than one large I/O. Previously, if the application did a
large read from a channel the data would be read from the file into the
buffer in a single large read, then copied from the buffer into the
application, requiring twice as much memory as the size of the read. A
similar behavior happened with writes.

With these changes even if the application does a large read the data
are read into the buffer and copied to the application in fixed-size
I/Os. The amount of memory is reduced to the size of the read plus the
size of the I/O. Writes are similar.

Changing this behavior required several changes and fixes to the channel
implementation. First, the channels are no longer marked when reading
or writing unstructered data such as strings and bytes. Marking a
channel causes all subsequent reads or writes to be buffered until the
channel is reverted or committed, requiring memory equal to the
subsequent reads or writes independent of the size of the I/Os to the
underlying file. There is little point in marking the channel anyway as
reverting a write is impossible and reverting a read of unstructured
data isn't very useful.

Second, there was an off-by-one error in the buffer implementation
necessitating qbuffer_iter_at_part_end.

Third, qio_memalign did not call chpl_mem_memalign so that memory
allocated in this way could not be tracked.

Fourth, the result of readBytesOrString was created through a copy, so
that in the case of a large read the memory requirements were three
times the size of the read. Using the swap operator avoids this copy.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>