Add nightly M1 Mac testing
Add comm=none release/example M1 testing. We want to increase the amount
of testing we do over time, but to start we just want to get some
testing going. Note that this adds a suppression for a known bug with
the fileIO primer on M1 (19218)

Part of Cray/chapel-private 2625

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>