Add open() and its siblings, and a test thereof.
This adds, creat(), open(), read(), write(), and close(), along with
O_CREAT and all the other mode bits.  It also adds a runtime wrapper for
open(), which we need because we can't (afaik) exactly represent the C
varargs behavior for POSIX open() directly in the Chapel module code.
(If this is untrue or becomes so at some point then we can simplify
then, but it's not much code and also isn't something that will arise in
many other places, so no big deal really.)

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>