Have build version print out for developers or non-zero BUILD_VERSIONs
It appears that in https://github.com/chapel-lang/chapel/pull/13191
I got my boolean logic mixed up while refactoring how we print
versions such that the build version would only print for official
releases if developer was set.  Prior to that PR, it was printed
for official releases if non-zero even if developer was un-set, and
this seems like a smarter default (such that we'd print 1.26.0 for
the first of any official releases and then 1.26.0.1 if the build
version was bumped).

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>