Remove slice; avoid bug
Removing the slice less for performance (it's only done 3 times) and
more because it seems silly to avoid passing in the bounds only to
look them right back up again within the routine.

Better than 'lo = bufflo' would be to pass 'lo' in by 'in' intent
but this triggers a copy propagation assertion in the compiler
for reasons I'm uncertain of.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>