Remove internal modules from the module index
The approach here is not particularly attractive: Since chpldoc can't
distinguish internal from non-internal modules, I'm using a name match
to filter out the internal modules.  The rest of the work in this
commit involved removing references to the internal modules (which are
appropriate anyway since they're not user-facing) and coming up with
other ways of preserving the similar hyperlinking behavior (e.g.,
introducing a `string` and `bytes` type tag for all of the numerous
string/bytes cross-references that we have to jump to).

While here, fixed some things that were currently broken in the docs.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>