Implement a better fix for an MCM issue in ofi msg-order-fence
Implement a better fix for 19575. In that PR we forced memory visibility
for all nodes before doing an AM to fix a MCM race. This was because a
fenced send did not appear to be fencing, but it turns out that we were
clearing the bitmask used to track outstanding operations for all nodes
instead of excluding the current one. That resulted in issuing a regular
non-fenced send. Fix that by only clearing pending ops for the nodes we
forced visibility for. This allows us to revert the previous workaround.

Part of Cray/chapel-private 3146

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>