Correct `CTypes` module usage.
I mis-coded the `use CTypes` in `OS.POSIX` and as a result also had to
`use CTypes` in several of the tests.  My intention had been that
`POSIX` should re-export the `CTypes` types, however.  Here, make the
`use` in POSIX public and thereby allow removing the ones in the tests.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>