Fix nightly memleak failure in DistributedMap
This PR does not fix the memleaks. It sets up a prediff
so that the test passes nightly testing, following the lead of:

    test/memory/ferguson/test1.chpl

plus sorting the output, which is needed to match .good reliably.

While there, reduce the problem size so that the test runs quickly.
If we want a stress test with higher problem size,
we should run it without --memLeaks.

Signed-off-by: Vassily Litvinov <vasslitvinov@users.noreply.github.com>