Skipif memleaks testing of DistributedMap
#19624 did not succeed making DistributedMap nightly testing pass
because various configurations, such as ASAN, cygwin, etc., have
different sets of leaks.

I believe that tracking the exact set of leaks exhibited by this test
is of little value, given that this code is just a sketch of an idea
and because it involves a privatized data type.

Therefore this PR reverts #19624 and instead adds a skipif file
to skip memleaks testing.

Signed-off-by: Vassily Litvinov <vasslitvinov@users.noreply.github.com>