Update and retire a future whose behavior is now supported
This future relates to wanting to refer to a 'struct foo' type within
Chapel as an extern record, which we didn't have support for in the
early days of interop.  Now we can rename such a case using "struct
foo" as the C name in the extern record declaration.  Update the test
to do so and retire the future and its skipif (no longer needed).

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>