Adjust some tests to use the `OS.POSIX` module instead of `Sys`.
Mostly these are just `getenv()` calls. `rename_extern_record` is an
outlier -- it uses `struct stat` in a test for renamed extern records,
and the change is mostly the addition of an explanatory comment.  The
`users/ferguson/sys/getenv/getenv.chpl` test is removed because it is
in effect replaced by `test/modules/OS/POSIX/getenv.chpl`.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>