Fix bitwise-bool warnings in the compiler
This adjusts a code pattern in our compiler that used to causes warnings
about `&` on booleans.

Details:

The code uses `&` instead of `&&` to continue checking even after
errors have been found, in order to report as many errors as possible
in one compilation. I thought about changing the individual functions
to return ints instead of booleans, however this solution results
in clearer code because it preserves the boolean-ness of the checking functions.

Signed-off-by: Vassily Litvinov <vasslitvinov@users.noreply.github.com>