Remove two internal compiler errors that were blocking valid statements
Before, we were halting inside the compiler when an import statement contained
multiple "this" or "super" but no other names.  `import super.super;` etc should
be valid, while `import super.this` is not really better than `import super;`
but shouldn't be a problem to support.  Both of these were blocked by two
INT_FATALs in the compiler, so remove those cases and add a warning when an
import statement ends with "super.this" or "this.this".

This isn't quite as extensive as I would like.  For one thing, it won't detect
something like `super.this.super` or `this.this.super` and generate the warnings
in those cases.  However, it will still allow them to compile and I don't think
will allow bad cases (though full testing is still needed).

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>