Fix bug with importing private submodules
The resolution of import/use statements was failing to consider whether the
symbol it was matching against was private to the module in which it was found
via a use/import chain.  This adjusts ResolveScope::followImportUseChains to
skip including symbols that are private.

I'm getting a bit of deja vu here, so this might not work.  Will check

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>