Adjust how new expressions are converted
Also add a .prediff for the 'multiple-modifiers' test which checks
for multiple modifiers e.g., 'owned borrowed shared C'. Today,
this is run during 'checkParsed()', which iterates over calls in
insertion order. We should migrate as much of this pass to dyno
as possible and use a tree-walk instead (at which point this prediff
can be removed).

Signed-off-by: David Longnecker <dlongnecke-cray@users.noreply.github.com>