Make job name configurable
Allow configuration of the name of the job created by the Chapel
launcher. If the CHPL_LAUNCHER_JOB_NAME environment variable is set it
is used as the job name. Otherwise, if CHPL_LAUNCHER_JOB_PREFIX is set
it is used as the prefix for the job name, else "CHPL-" is used as the
prefix.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>