Resolves #18565
Improve the ordering of non-matching resolution candidates

This changes the order and sometimes the number of resolution candidates
that the compiler prints upon an "unresolved call" error message as
"note: this candidate did not match" and "note: other candidates are".

* Print all the candidates if there are 4 or fewer.
  There should no longer be "note: there is also 1 other candidate".
  Instead the compiler just shows that 1 other candidate.

* Print the candidates with the same method-ness as the unresolved call
  BEFORE the other-methodness candidates.
  Ex. if the call is a method call, print the candidates that are methods
  before those that are standalone functions.
  Note that this is not perfect in the current compiler, which treats
  as "method calls" some calls written without an explicit receiver.

* Towards #18565, print the candidate methods with the same receiver type
  as the call (if it is a method call) BEFORE those with different
  receiver types.

The last bullet is implemented by adding another "reason" for rejecting
a resolution candidate: `RESOLUTION_CANDIDATE_DIFFERENT_RECEIVER_TYPES`
which is ordered after most other reasons.

Potential future improvement: order the candidates whose receiver type
is a supertype of the receiver type of the call BEFORE the other
different-receiver-type candidates.

As before, the compiler explains the first non-matching candidate in detail
with "note: this candidate did not match: ..." and "note: because ...".
It can be a different candidate now than it would have been before this PR.

While there:
* Add .bad: `test/classes/nilability/cycle-generic.bad`
* Add .future for #19697: `test/errhandling/parallel/taskerrors-write.chpl`

Signed-off-by: Vassily Litvinov <vasslitvinov@users.noreply.github.com>