Adjust tests to pass on macOS
I ran into these while testing an unrelated change.  Here, adjust several
tests so that they pass on macOS.

Rename `extern/ferguson/ctypes.chpl` to `t-ctypes.chpl`, because the
case-insensitive nature of the macOS filesystem meant that this file was
being confused as the intended source of the `CTypes` module for `use`
statements in this directory.

In `test/llvm/alignment/check-align16.prediff`, remove the `sed` '-s'
option.  macOS's `sed` doesn't like it, and it was spurious anyway.
Also, change `check-align16-field1.chpl` to use the same size-specific
type specifiers in its `printf()` format string as `check-align16.chpl`
already does.  Otherwise, the target compiler gripes about a size
mismatch between the format specifier and the thing to be printed.

Finally, in `parallel/taskPool/figueroa/ManyThreads.prediff`, change from
a '+' operator to a '*' operator to get around a difference between macOS
and Linux `sed`.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>