Eureka feature proposal
This is a sketch implementation of support for eurekas, #12700.
See also design questions in https://github.com/chapel-lang/chapel/issues/12700#issuecomment-1106837084
and an alternative approach in #12815.

User code using this feature looks like this:

    use Eureka;
    var e: eureka;

    forall ..... {
      e.check(); // this could be inserted automatically by the compiler
      doWork();
      if checkForEureka() then
	e.notify()
    }

Within the loop body `e` carries a task-private flag.

`e.notify()` sets this flag for all tasks created by the forall loop, then (or concurrently) throws a EurekaNotified error.

`e.check()` checks if the current task's flag has been set. If so, it throws a EurekaChecked error. Otherwise, it is a cheap operation -- read and branch on an atomic bool.

The implementation is intentionally simplistic for ease of understanding. Here are steps to productize it, starting with the most immediate concerns:

* Beef up the implementation to prevent race conditions.
  See todo comments in the code.
  Or switch to another implemenatation.

* Tune performance, esp. for multilocale programs.

* When using this feature for task-parallel constructs,
  do not require explicit in-intent on the eureka variable.
  For that, account for FLAG_DEFAULT_TFI_IN in task intents.

* Have `e.check()` be inserted automatically by the compiler.

* (optional) Make `eureka` customizable. Ex. allow argument(s) to `e.notify()`
  that are passed into the exceptions that get thrown.

* Allow the use of this feature via `break` instead of `e.notify()`
  and error handling. This would entail automatic insertion of
  `var e: eureka` and error handling; ex. chpl_forall_error() would drop
  EurekaNotified and EurekaChecked errors.

Note: the demo test leaks an instance of the EurekaImpl class.
The PR follows the lead of test/memory/ferguson/test1.chpl
to ensure that nightly testing passes.
I believe this leak is due to the same cause as the one from #19479:
`test/errhandling/parallel/forall-calls-throwing-fn2.chpl`

Signed-off-by: Vassily Litvinov <vasslitvinov@users.noreply.github.com>