Fix an internal error with generic nilable class
This PR:
* resolves the crash for the code in the OP of #14148
* cleans up the error reported upon a variant of it
* files a .future for the code in the first comment of #14148

DETAILS

Given

```chpl
proc p(): MyClass {
  if ... then  return nil;
  else         return new MyClass();
}
```

The compiler inserts a cast `nil: MyClass` for the first return
because it does not yet know that the returned type is the `owned` flavor
of MyClass. This used to crash the compiler. Now this causes a user-friendly
error message.

The same error message is produced if the order of the returns is switched.

We could probably teach the compiler to infer the ownership of the return
type and use that to cast `nil` to. However, that may be a significant effort.
For now, punt and have the user adjust the code.

The expressions `nil:owned` and `nil:shared` used to cause obscure error
messages, now they cause the compiler to state clearly that it is illegal.
We could potentially introduced the "owned nil" and "shared nil" types
so that these casts succeed. I am leaving that for another day.

Signed-off-by: Vassily Litvinov <vasslitvinov@users.noreply.github.com>