Reflecting review, adjust the `select()` test to prevent deadlock.
The two tasks in `select.chpl` were non-cooperating, so in the event
they ended up scheduled on a single execution vehicle thread the test
would have deadlocked.  Here, add task yields to make them cooperative.
This is sufficient to prevent deadlock with qthreads tasking, though not
with fifo tasking.  Therefore in addition, add a skipif so that this
test only runs with qthreads tasking.

While here, add a skipif so that the `pipe.chpl` test is also not run
except with qthreads tasking.  That one already uses a `single` var to
order its output, which as a side effect will prevent the deadlocks
described above, but it needs the skipif.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>