Adjust scope resolve to avoid recursive queries
Recursive queries as implemented today present some challenge for the
query system and we have some problems with incremental compilation when
they are used. In particular when the incremental framework recomputes a
query (to check a dependency) it does not currently handle the
possibility of a recursive query. At the same time, the current approach
can have partial results and these are mutated partway though - so there
might be a case where the partial result used by some other query is
different in one case or another (for example, if some, but not all, of
the dependent queries could be reused, the partial data might or might
not appoear visible at all).

There are currently 3 locations that use recursive queries today:
 * scope resolution of what module is use/imported
 * resolving the module-level statements in a module (not including procs)
 * resolving the fields of recursive types

This commit changes the first case to no longer use a recursive query.
Instead, changes approach to having a query to resolve an individual
use/import statement. That query takes in a "path" of which use/import
statements are currently being resolved, so that the resolution of the
use/import statements can consider only the symbols available up to that
point.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>