Fix problem with dyno builds on Mac OS X with LLVM
We were observing problems with 'make test-dyno' when CHPL_LLVM=system
on Mac OS X with Homebrew cmake and llvm.

The cmake build tries to gather the command line arguments
from chplenv but sometimes these include -isysroot, but cmake wants to
have its own variable to set that. This PR just adjusts the Makefiles to
configure cmake's sdkroot to match whatever -isysroot our chplenv scripts
have figured out. That way, if cmake adds a flag for it in addition, it
is OK, because everything is consistent.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>