Reduce CHPL_HOME assumptions when running 'make docs'
This makes some changes to 'make docs' to reduce assumptions, such as:
* that CHPL_HOME is set
* that CHPL_HOME / pwd is not a symbolic link

Specifically, the motivation for this is that some of our nightly test
configurations that use a symbolic link for CHPL_HOME were failing
when building 'make docs' because the logic I added in #19584 to
distinguish internal modules from others was failing due to "real"
paths being used by the compiler and symbolic ones by our Makefile.
Here, the fix is to add a '--home' developer flag to 'chpldoc' and use
it when we're invoking it such that we're guaranteed that its notion
of CHPL_HOME matches the one we use when passing in the internal
modules.

In more detail:

Part of the work in adding the module index sidebar was to prevent internal
modules from getting added to the module index since:

(a) users aren't really meant to know their names nor that they even
    exist, and

(b) we're working on rolling the ones that are important into the spec
    anyway

However, since our 'chpldoc' invocations just list the modules to document
on the command-line, there wasn't a pre-existing way to distinguish internal
modules from standard/package/user modules.  As a result, I added some logic
to the chpldoc branches of the compiler to see whether a given module lives
in $CHPL_HOME/modules/internal, and to tag it as internal if it is.  Then I
changed our Makefile to provide the internal modules to 'chpldoc' using
their absolute paths so that this comparison would pass (at least in the
tests I ran).

What's happening in this situation where CHPL_HOME is a symbolic link is
that the compiler is inferring CHPL_HOME to be the "real" path, but our
Makefiles are providing paths relative to the symbolic link, so the compiler
doesn't correctly recognize them as being internal using its strcmp() test.

The fix I'm taking here is to have 'chpldoc' support a '--home' flag like
'chpl' does, intended for developers' use, and to have our Makefile pass in
what it considers CHPL_HOME to be so that the path is guaranteed to match
the one it uses for specifying the internal modules.

Some other approaches considered in the original PR, but passed on, which
could also be used instead of this approach:

* we could pass relative paths as we have and just cue off of
  'modules/internal': this would work, but felt unfortunate if an end-user
  just happened to use that same path

* make 'chpldoc' argument parsing support an --internal or --noindex flag
  that could be supplied to indicate that the following modules were
  internal (or maybe all of them, breaking the current single 'chpldoc'
  invocation into two).  This felt more fiddly than the approach I
  originally took, and I like having chpldoc be able to distinguish
  internal modules, which may come in handy in other ways.~

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>