Move prototype of chplHome from driver.h to commonFlags.h
In reviewing #19753, Lydia pointed out that commonFlags.h now referred
to 'setHome' yet didn't #include driver.h, which is where that routine
is prototyped.  This wasn't an issue because each file that #includes
commonFlags.h also #includes driver.h, but it's obviously a bit
sketchy (and didn't get addressed in that PR because I was hoping to
fix the broken build last night).

Here, I'm moving the prototype of setHome to commonFlags.h to fix
this issue, in the section of other prototypes that are used to
set things in the common flags section.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>