Revert changes to isBracketLoopMaybeArrayType
In a previous commit, we were allowing [] loops with if-expr bodies
to be marked as "maybe array types". I stopped doing that, thinking
that they should all be handled by `convertArrayType`. Hilariously,
this caused the one test that Ahmad had called out in a comment to
start failing again. After re-reading the grammar, revert my revert.

Signed-off-by: David Longnecker <dlongnecke-cray@users.noreply.github.com>