Resolve loop index variables in For loops
Updates the Resolver visitor to capture handling of For loops in order
to resolve loop index variables. For iterand expressions that are not
direct iterator calls, this visitor will attempt to resolve a ".these"
iterator call on the iterand. If successful the resolved method will be
stored in the iterand's associatedFns.

If a ".these" call could not be resolved the compiler will issue an
error:
  "unable to iterate over values of type <type>"

If resolution of the iterand failed the loop index variable will have
the ErroneousType.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>