Deprecate VectorizingIterator module
In the previous release, we deprecated the vectorizingIterator()s
themselves, which felt like a precondition to deprecating the module
itself given that it was an auto-included module.  Here, we undo
the logic to deprecate the iterator, and just deprecate the module
as a whole, removing it from the list of auto-included modules.
That said, it can still be manually used as ever.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>