Limit timeout for this test
This future test switches between getting a nil dereference error and
timing out in its valgrind runs for reasons that aren't clear (task
scheduling?  base address layout?), and I expect it would be hard to
make the failure more predictable without simply fixing the bug.  The
change in behavior isn't problematic except that the default valgrind
timeout is quite high to give time for real programs to execute.
Here, I'm limiting the timeout to 10 seconds to avoid wasting time
because the program should either pass or fail in that amount of time.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>